test(parser): require split wasm PMX parity
